+++ traffic_portal/app/src/common/api/TypeService.js
@@ -83,6 +83,31 @@ var TypeService = function($http, ENV, locationUtils, messageModel) {
         );
     };
 
+    this.queueServerUpdates = function(cdnID, typeName) {
+        return $http.post(ENV.api['root'] + 'cdns/' + cdnID +'/queue_update?type=' + typeName, {action: "queue"}).then(
+            function(result) {
+                messageModel.setMessages([{level: 'success', text: 'Queued server updates by type'}], false);
+                return result;
+            },
+            function(err) {
+                messageModel.setMessages(err.data.alerts, false);
+                throw err;
+            }
+        );
+    };
+
+    this.clearServerUpdates = function(cdnID, typeName) {
+        return $http.post(ENV.api['root'] + 'cdns/' + cdnID + '/queue_update?type=' + typeName, {action: "dequeue"}).then(
+            function(result) {
+                messageModel.setMessages([{level: 'success', text: 'Cleared server updates by type'}], false);
+                return result;
+            },
+            function(err) {
+                messageModel.setMessages(err.data.alerts, false);
+                throw err;
+            }
+        );
+    };
 };
 
 TypeService.$inject = ['$http', 'ENV', 'locationUtils', 'messageModel'];